requirements.txt
asgiref==3.4.1
Django==3.2.11
django-cors-headers==3.11.0
djangorestframework==3.13.1
gunicorn==20.1.0
install==1.3.5
pytz==2021.3
sqlparse==0.4.2
whitenoise==5.3.0
